added GET clb_ICclub endpoint
