+++ kbsb/interclubs/md_interclubs.py
@@ -10,6 +10,40 @@ from typing import Dict, Any, List, Optional
 from xml.dom.expatbuilder import DOCUMENT_NODE
 from pydantic import BaseModel
 from typing import Literal
+from reddevil.core.dbbase import DbBase
+
+# DB classes
+
+
+class DbICSeries(DbBase):
+    COLLECTION = "interclub2324series"
+    DOCUMENTTYPE = "InterclubSeries"
+    VERSION = 1
+    IDGENERATOR = "uuid"
+
+
+class DbICVenue(DbBase):
+    COLLECTION = "interclub2324venues"
+    DOCUMENTTYPE = "InterclubVenues"
+    VERSION = 1
+    IDGENERATOR = "uuid"
+    HISTORY = True
+
+
+class DbICClub(DbBase):
+    COLLECTION = "interclub2324club"
+    DOCUMENTTYPE = "ICClub"
+    VERSION = 1
+    IDGENERATOR = "uuid"
+    HISTORY = True
+
+
+class DbICEnrollment(DbBase):
+    COLLECTION = "interclub2324enrollment"
+    DOCUMENTTYPE = "InterclubEnrollment"
+    VERSION = 1
+    IDGENERATOR = "uuid"
+    HISTORY = True
 
 
 # interclubclub
@@ -45,17 +79,17 @@ class ICPlayer(BaseModel):
     first_name: str
     idnumber: int
     idcluborig: int  # the club the player belongs to in signaletique
-    idclubvist: int  # the club the playr is playing if he plays elsewhere
+    idclubvist: int  # the club the player is playing if he plays elsewhere
     last_name: str
     natrating: int
     nature: Literal[
         "assigned",
+        "unassigned",
         "requestedout",
         "requestedin",
         "comfirmedin",
         "confirmedout",
-        "unassigned",
-        "absent",
+        "locked",
     ]
     transfer: ICTransfer | None = None
